tools/refresh.sh:skip config check on debug configs

+++ tools/refresh.sh
@@ -207,6 +207,12 @@ for CONFIG in ${CONFIGS}; do
     fi
   fi
 
+  # skip refresh if defconfig contains `#include`
+  if grep -q "#include" $DEFCONFIG; then
+    echo "Note: skipping refresh for debug defconfig."
+    exit 0
+  fi
+
   # Copy the .config and Make.defs to the toplevel directory
 
   rm -f SAVEconfig
